Replace pyyaml with ruamel.yaml
The `pyyaml` library doesn't preserve list orders but there is a need to
know which source is the first one. Therefore it has been replaced with
`ruamel.yaml` which preserves order by default.

Custom classes are used in `ruamel.yaml`. Therefore tests if a type is a
dict or list was replaced with `isinstance`.