Don't prefix versions with `v`
This change is motivated by the Anaconda package repository's preference
logic. Because of that logic uploaded packages that have a version
starting with a number are always preferred over those that have a
version starting with `v`.

Therefore if the package X previously had version like `1.2.3` and the
newer version is `v1.5.8` then the `conda install X` command will
install the `X-1.2.3` package. This is the case of many packages
currently built with `conda-build-prepare`.