CI: Fix `build/number: None` in the test recipe
It seems that `conda-build v3.20.*` didn't care about these environment
variables being missing in the CI which results in substituting `None`
there but rendering with the lately released `v3.21.*` fails due to
such an error:

```
ValueError: Build number was invalid value 'None'. Must be an integer.
```

The values have been simply given a default value to fix the issue.