Always put channels from `--channels` on top of package's `condarc`
The package's `condarc` is always used to either:
* add a backup channel for missing packages,
* put some channel(s) in front of the `defaults` channel.

In both of these situations, channels passed with the `--channels`
switch should remain being the top channels during building.

The `conda config ...` and `condarc` blocks have to be reversed because
the `condarc` created with all the channels seems to influence the
`conda config` channel setting significantly.