Print stdout if check_output's subprocess fails
`stdout` is captured while calling commands with `check_output`. The
default exception doesn't even mention what the command has printed to
`stdout` even though it is available in the CalledProcessError instance.

This is especially important for Conda as it sometimes prints errors to
STDOUT. Not printing it can slow down fixing significantly. It's
definitely true in case of a recipe source patching problems.